+++ src/Codeception/Codecept.php
@@ -130,6 +130,7 @@ class Codecept
         $this->dispatcher->addSubscriber(new Subscriber\BeforeAfterClass());
         $this->dispatcher->addSubscriber(new Subscriber\AutoRebuild());
         $this->dispatcher->addSubscriber(new Subscriber\Logger());
+        $this->dispatcher->addSubscriber(new Subscriber\Bootstrap());
 
         // optional
         if (!$this->options['silent'])  $this->dispatcher->addSubscriber(new Subscriber\Console($this->options));