+++ tests/cli/OrderCest.php
@@ -1,14 +1,14 @@
 <?php
 
-class OrderCest {
-
+class OrderCest
+{
     public function checkOneFile(CliGuy $I)
     {
         $I->amInPath('tests/data/sandbox');
         $I->executeCommand('run order LoadingOrderCept.php');
         $I->expect('global bootstrap, initialization, beforeSuite, before, bootstrap(B), test(T), after, afterSuite');
         $I->seeFileFound('order.txt','tests/_log');
-        $I->seeFileContentsEqual("BI([BST])");
+        $I->seeFileContentsEqual("BI(B[ST])");
     }
 
     public function checkForFails(CliGuy $I)
@@ -17,7 +17,7 @@ class OrderCest {
         $I->executeCommand('run order FailedCept.php --no-exit');
         $I->seeFileFound('order.txt','tests/_log');
         $I->expect('global bootstrap, initialization, beforeSuite, before, bootstrap, test, fail, after, afterSuite');
-        $I->seeFileContentsEqual("BI([BSTF])");
+        $I->seeFileContentsEqual("BI(B[STF])");
     }
 
 
@@ -26,7 +26,7 @@ class OrderCest {
         $I->amInPath('tests/data/sandbox');
         $I->executeCommand('run order --no-exit --group simple');
         $I->seeFileFound('order.txt','tests/_log');
-        $I->seeFileContentsEqual("BI({[BST][BSTF][BST])}");
+        $I->seeFileContentsEqual("BI({B[ST][STF][ST])}");
     }
 
     public function checkCestOrder(CliGuy $I)
@@ -34,7 +34,7 @@ class OrderCest {
         $I->amInPath('tests/data/sandbox');
         $I->executeCommand('run tests/order/ReorderCest.php --no-exit');
         $I->seeFileFound('order.txt','tests/_log');
-        $I->seeFileContentsEqual("BI([B0123456])");
+        $I->seeFileContentsEqual("BI(B[0123456])");
     }
 
     public function checkCodeceptionTest(CliGuy $I)