+++ tests/data/claypit/tests/dummy/FileExistsCept.php
@@ -1,4 +1,6 @@
 <?php
+require '_bootstrap.php';
+
 $I = new DumbGuy($scenario);
 $I->wantTo('check config exists');
 $I->seeFileFound($codeception);