+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+#### 2.0.0-beta 03/18/2014
+
+* bootstrap files are loaded before suite only
+* extension can reconfigure global config
+
+
 #### 2.0.0-beta 03/17/2014
 
 * removed RefactorAddNamespace and Analyze commands